refactor: sevice 패키지 service로 오타 수정
