+++ src/test/java/dandi/dandi/pushnotification/application/service/WeatherPushNotificationSchedulerTest.java
@@ -12,8 +12,7 @@ import dandi.dandi.member.application.port.out.MemberPersistencePort;
 import dandi.dandi.member.domain.Location;
 import dandi.dandi.pushnotification.application.port.out.persistence.PushNotificationPersistencePort;
 import dandi.dandi.pushnotification.application.port.out.webpush.WebPushManager;
-import dandi.dandi.pushnotification.application.sevice.WeatherPushNotificationScheduler;
-import dandi.dandi.pushnotification.application.sevice.message.WeatherPushNotificationMessageGenerator;
+import dandi.dandi.pushnotification.application.service.message.WeatherPushNotificationMessageGenerator;
 import dandi.dandi.pushnotification.domain.PushNotification;
 import dandi.dandi.pushnotification.domain.PushNotificationTime;
 import dandi.dandi.weather.application.port.out.WeatherForecast;
